requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
langchain-community
python-dotenv
tiktoken
langchain-google-genai
google-generativeai